Remove faulty metadata parsing and use only file names
The metadata parsing was unreliable, just keep it simple and name each
podcast entry based on its filename.

Also add some more logging and log the response code sent to the client.
Replace the previous use of http.FileServer with a simpler
implementation that only serves the files included in the podcast feed.